- Moving resolvers to his own package folder. - Fixes in the sign in usecase.
