+++ cmd/server/run.go
@@ -23,6 +23,7 @@ import (
 	"github.com/daniarmas/chat/internal/datasource/stream"
 	"github.com/daniarmas/chat/internal/delivery/graph"
 	"github.com/daniarmas/chat/internal/delivery/graph/middleware"
+	"github.com/daniarmas/chat/internal/delivery/graph/resolver"
 	"github.com/daniarmas/chat/internal/repository"
 	"github.com/daniarmas/chat/internal/usecases"
 	ownredis "github.com/daniarmas/chat/pkg/own-redis"
@@ -145,7 +146,7 @@ to quickly create a Cobra application.`,
 
 		router.Use(middleware.AuthorizationMiddleware(jwtDs, accessTokenRepo))
 
-		srv := handler.NewDefaultServer(graph.NewExecutableSchema(graph.Config{Resolvers: &graph.Resolver{AuthUsecase: authUsecase, MessageUsecase: messageUsecase, ChatUsecase: chatUsecase}}))
+		srv := handler.NewDefaultServer(graph.NewExecutableSchema(graph.Config{Resolvers: &resolver.Resolver{AuthUsecase: authUsecase, MessageUsecase: messageUsecase, ChatUsecase: chatUsecase}}))
 
 		srv.AddTransport(transport.POST{})
 		srv.AddTransport(transport.Websocket{
