+++ gqlgen.yml
@@ -20,8 +20,8 @@ model:
 # Where should the resolver implementations go?
 resolver:
   layout: follow-schema
-  dir: internal/delivery/graph
-  package: graph
+  dir: internal/delivery/graph/resolver
+  package: resolver
   filename_template: "{name}.resolvers.go"
   # Optional: turn on to not generate template comments above resolvers
 # omit_template_comment: false
